minecraft_voxelizer/requirements.txt
torch
torchvision
transformers
numpy
pillow
scipy
matplotlib
open3d
opencv-python